+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- next-header -->
 ## [Unreleased] <!-- ReleaseDate -->
 
+- add optional postcard MaxSize
+
 ## [2.2.0]  - 2023-08-24
 
 - enums: Eq + PartialEq